Rework the parser modules
