+++ src/aiida_abacus/parsers/__init__.py
@@ -1,0 +1,4 @@
+from .abacus import AbacusParser
+from .raw_parsers import AbacusRawParser
+
+__all__ = ["AbacusParser", "AbacusRawParser"]
