+++ src/aiida_abacus/parsers/abacus.py
@@ -1,0 +1,130 @@
+"""
+Parsers provided by aiida_abacus.
+
+Register parsers via the "aiida.parsers" entry point in setup.json.
+"""
+
+import re
+from typing import TextIO
+
+from aiida import orm
+from aiida.common import exceptions
+from aiida.parsers.parser import Parser
+from aiida.plugins import CalculationFactory
+from ase.io.cif import read_cif
+
+from ..common import make_retrieve_list
+from .raw_parsers import AbacusRawParser
+
+AbacusCalculation = CalculationFactory("abacus.abacus")
+
+
+class AbacusParser(Parser):
+    """
+    Parser class for parsing output of calculation.
+    """
+
+    def __init__(self, node):
+        """
+        Initialize Parser instance
+
+        Checks that the ProcessNode being passed was produced by a AbacusCalculation.
+
+        :param node: ProcessNode of calculation
+        :param type node: :class:`aiida.orm.nodes.process.process.ProcessNode`
+        """
+        super().__init__(node)
+        if not issubclass(node.process_class, AbacusCalculation):
+            raise exceptions.ParsingError("Can only parse AbacusCalculation")
+
+    def parse(self, **kwargs):
+        """
+        Parse outputs, store results in database.
+
+        :returns: an exit code, if parsing fails (or nothing if parsing succeeds)
+        """
+        output_folder = self.retrieved
+        settings = {} if "settings" not in self.node.inputs else self.node.inputs.settings
+        expected_files = make_retrieve_list(self.node.inputs.parameters, settings, AbacusCalculation._OUTPUT_SUFFIX)
+        # Add the STDOUT diversion
+        expected_files.append(AbacusCalculation._ABACUS_OUTPUT)
+        run_type = self.node.inputs.parameters["input"].get("calculation", "scf")
+
+        # Check if the files are retrieved
+        missing = []
+        for name in expected_files:
+            try:
+                output_folder.get_object(name)
+            except FileNotFoundError:
+                missing.append(name)
+
+        if missing:
+            self.logger.warning(f"The following expected files are missing: {missing}")
+
+        # Parse the calculation task output file
+        main_log = next(filter(lambda x: "running_" in x, expected_files))
+        misc_results = {}
+        with output_folder.open(main_log, "r") as fhandle:
+            parser = AbacusRawParser(fhandle)
+            misc_results.update(parser.parse())
+        misc_node = orm.Dict(dict=misc_results)
+
+        # Parse the structure output
+        fname = next(filter(lambda x: "STRU.cif" in x, expected_files))
+        # TODO: there could be other types that should have a output structure
+        if run_type in ["relax", "cell-relax", "md"]:
+            with output_folder.open(fname, "r") as fhandle:
+                atoms = read_cif(fhandle)
+                self.out("structure", orm.StructureData(ase=atoms))
+
+        # Parse the calculation raw parameters
+        if "settings" in self.node.inputs and self.node.inputs.settings.get("include_internal_parameters", False):
+            fname = next(filter(lambda x: x.endswith("INPUT"), expected_files))
+            with output_folder.open(fname, "r") as fhandle:
+                self.out("internal_parameters", read_internal_parameters(fhandle))
+
+        # Parse the KPOINTS actually used
+        if "settings" in self.node.inputs and self.node.inputs.settings.get("include_kpoints", False):
+            fname = next(filter(lambda x: x.endswith("kpoints"), expected_files))
+            with output_folder.open(fname, "r") as fhandle:
+                coords, weights = read_kpoints_output_file(fhandle)
+                node = orm.KpointsData()
+                node.set_kpoints(coords, weights=weights)
+                # Set the cell based on the  INPUT structure
+                node.set_cell_from_structure(self.node.inputs.structure)
+            self.out("kpoints", node)
+
+        # Define the output nodes
+        self.out("misc", misc_node)
+
+
+def read_internal_parameters(fhandle: TextIO):
+    """Read the internal parameters"""
+    fhandle.readline()
+    out_dict = {}
+    for line in fhandle:
+        if line.startswith("#"):
+            continue
+        # Remove the trialing # comments
+        match = re.match(r"^(.+) #.*$", line)
+        if match:
+            tokens = match.group(1).split()
+            out_dict[tokens[0]] = tokens[1]
+    return out_dict
+
+
+def read_kpoints_output_file(fhandle: TextIO):
+    """Read the output kpoints file"""
+
+    line = fhandle.readline()
+    nkpts = int(line.strip().split()[-1])
+    assert fhandle.readline().startswith("K-POINTS DIRECT COORDINATES")
+    fhandle.readline()
+    points = []
+    weights = []
+    for i in range(nkpts):
+        tokens = fhandle.readline().strip().split()
+        points.append([float(tokens[i]) for i in range(1, 4)])
+        weights.append(float(tokens[4]))
+
+    return points, weights
