+++ tests/test_data/band_Al_pw/README.md
@@ -1,0 +1,3 @@
+# Example band structure calculation data
+Data taken from `band/Al_pw`
+Increased the number of spins = 2
